+++ backend/src/repositories/read/FeedReadDao.ts
@@ -17,6 +17,9 @@ import { Errors } from "@/utils/errors";
 import { logger } from "@/utils/winston";
 import {
   ACTIVE_POST_FILTER,
+  getStandardLookups,
+  getStandardProjectionFields,
+  normalizeObjectId,
   withActivePostFilter,
 } from "@/repositories/post-pipeline.helpers";
 
@@ -35,26 +38,6 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
     super(model);
   }
 
-  protected normalizeObjectId(
-    id: unknown,
-    field: string,
-  ): mongoose.Types.ObjectId {
-    if (id instanceof mongoose.Types.ObjectId) {
-      return id;
-    }
-
-    if (typeof id !== "string" || id.length === 0) {
-      throw Errors.validation(`${field} is not a valid ObjectId`);
-    }
-
-    try {
-      return new mongoose.Types.ObjectId(id);
-    } catch {
-      throw Errors.validation(`${field} is not a valid ObjectId`);
-    }
-  }
-
-  // Extracted logic starts here
   private readonly tagIdCacheStore = new Map<
     string,
     { ids: mongoose.Types.ObjectId[]; expiresAt: number }
@@ -115,7 +98,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
       let results: ProjectedFeedPost[] = [];
 
       const feedProjection = {
-        ...this.getStandardProjectionFields(),
+        ...getStandardProjectionFields(),
         _id: 1,
       };
 
@@ -131,7 +114,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
         pipeline.push(
           { $sort: { createdAt: -1, _id: -1 } },
           { $limit: limit + 1 },
-          ...this.getStandardLookups(),
+          ...getStandardLookups(),
           {
             $addFields: {
               tagNames: {
@@ -166,7 +149,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
           // If we ALREADY fetched some personalized posts on this CURRENT page, we should theoretically offset the backfill.
           // However, to keep it simple, we just start backfill from scratch and cursor tracking will remember where we are based on the Last Backfill Post.
           { $limit: needed },
-          ...this.getStandardLookups(),
+          ...getStandardLookups(),
           { $addFields: { isPersonalized: false } },
           { $project: feedProjection },
         ];
@@ -199,7 +182,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
         backfillPipeline.push(
           { $sort: { createdAt: -1, _id: -1 } },
           { $limit: limit + 1 },
-          ...this.getStandardLookups(),
+          ...getStandardLookups(),
           { $addFields: { isPersonalized: false } },
           { $project: feedProjection },
         );
@@ -235,117 +218,6 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
       throw Errors.database(
         (error instanceof Error ? error.message : String(error)) ??
           "failed to generate cursor feed",
-      );
-    }
-  }
-
-  /**
-   * Generates a ranked feed using weighted scoring (recency + popularity + tag match)
-   * @deprecated Use getRankedFeedWithCursor for better deep pagination performance
-   * The cursor-based variant avoids skip() overhead and scales better for infinite scroll
-   * @pattern Weighted Ranking Algorithm
-   * @strategy Time-decay scoring with logarithmic dampening for engagement metrics
-   * @complexity O(N log N) aggregation + O(skip) scan cost on skip-based pagination
-   * @note Limited to recent posts (90 days) to optimize performance and relevance
-   * @param favoriteTags - Array of tag names for tag-match scoring
-   * @param limit - Number of posts per page
-   * @param skip - Number of posts to skip (use cursor pagination for deep pages)
-   * @returns {Promise<PaginationResult<FeedPost>>} Ranked posts with computed scores
-   * @throws {DatabaseError} if score computation or aggregation fails
-   */
-
-  async getRankedFeed(
-    favoriteTags: string[],
-    limit: number,
-    skip: number,
-  ): Promise<PaginationResult<FeedPost>> {
-    try {
-      const weights = { recency: 0.5, popularity: 0.3, tagMatch: 0.2 };
-      const favoriteTagIds = await this.loadFavoriteTagIds(favoriteTags);
-      const hasTagPreferences = favoriteTagIds.length > 0;
-
-      // filter to recent posts to avoid full collection scan
-      const recentThresholdDays = 90;
-      const sinceDate = new Date(
-        Date.now() - recentThresholdDays * 24 * 60 * 60 * 1000,
-      );
-
-      const pipeline: PipelineStage[] = [
-        { $match: withActivePostFilter({ createdAt: { $gte: sinceDate } }) },
-        // compute ranking scores before $lookup to sort/paginate early
-        {
-          $addFields: {
-            recencyScore: {
-              $divide: [
-                1,
-                {
-                  $add: [
-                    1,
-                    {
-                      $divide: [
-                        { $subtract: [new Date(), "$createdAt"] },
-                        1000 * 60 * 60 * 24,
-                      ],
-                    },
-                  ],
-                },
-              ],
-            },
-            popularityScore: {
-              $ln: {
-                $add: [{ $max: [0, { $ifNull: ["$likesCount", 0] }] }, 1],
-              },
-            },
-            tagMatchScore: hasTagPreferences
-              ? { $size: { $setIntersection: ["$tags", favoriteTagIds] } }
-              : 0,
-          },
-        },
-        {
-          $addFields: {
-            rankScore: {
-              $add: [
-                { $multiply: ["$recencyScore", weights.recency] },
-                { $multiply: ["$popularityScore", weights.popularity] },
-                { $multiply: ["$tagMatchScore", weights.tagMatch] },
-              ],
-            },
-          },
-        },
-        { $sort: { rankScore: -1, _id: -1 } },
-        { $skip: skip },
-        { $limit: limit },
-        // now do expensive $lookups only on the paginated result set
-        ...this.getStandardLookups(),
-        {
-          $project: {
-            ...this.getStandardProjectionFields(),
-            viewsCount: { $ifNull: ["$viewsCount", 0] },
-            rankScore: 1,
-          },
-        },
-      ];
-      const [results, total] = await Promise.all([
-        this.model.aggregate(pipeline).exec(),
-        this.model.countDocuments(
-          withActivePostFilter({ createdAt: { $gte: sinceDate } }),
-        ),
-      ]);
-
-      logger.debug("Ranked feed generated", {
-        event: "feed.ranked.generated",
-        resultCount: results.length,
-        limit,
-        skip,
-      });
-
-      const totalPages = Math.ceil(total / limit);
-      const currentPage = Math.floor(skip / limit) + 1;
-      return { data: results, total, page: currentPage, limit, totalPages };
-    } catch (error: unknown) {
-      throw Errors.database(
-        (error instanceof Error ? error.message : String(error)) ??
-          "failed to build ranked feed",
       );
     }
   }
@@ -440,10 +312,10 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
         { $skip: skip },
         { $limit: limit },
         // now do expensive $lookups only on the paginated result set
-        ...this.getStandardLookups(),
+        ...getStandardLookups(),
         {
           $project: {
-            ...this.getStandardProjectionFields(),
+            ...getStandardProjectionFields(),
             viewsCount: { $ifNull: ["$viewsCount", 0] },
             trendScore: 1,
           },
@@ -494,10 +366,10 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
         { $sort: { createdAt: -1, _id: -1 } },
         { $skip: skip },
         { $limit: limit },
-        ...this.getStandardLookups(),
+        ...getStandardLookups(),
         {
           $project: {
-            ...this.getStandardProjectionFields(),
+            ...getStandardProjectionFields(),
             viewsCount: { $ifNull: ["$viewsCount", 0] },
           },
         },
@@ -680,10 +552,10 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
           : []),
         { $sort: { createdAt: sortDirection, _id: sortDirection } },
         { $limit: limit + 1 },
-        ...this.getStandardLookups(),
+        ...getStandardLookups(),
         {
           $project: {
-            ...this.getStandardProjectionFields(),
+            ...getStandardProjectionFields(),
             _id: 1,
             createdAt: 1,
             viewsCount: { $ifNull: ["$viewsCount", 0] },
@@ -805,7 +677,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
       }
 
       const sortDirection = direction === "forward" ? -1 : 1;
-      const feedProjection = this.getStandardProjectionFields();
+      const feedProjection = getStandardProjectionFields();
 
       const pipeline: PipelineStage[] = [
         {
@@ -863,7 +735,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
         { $sort: { trendScore: sortDirection, _id: sortDirection } },
         { $limit: limit + 1 },
         // populate relationships only on paginated results
-        ...this.getStandardLookups(),
+        ...getStandardLookups(),
         {
           $project: {
             ...feedProjection,
@@ -919,14 +791,8 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
 
   /**
    * Cursor-based pagination for ranked feed
-   * @description optimized for deep pagination on ranked posts
-   * @pattern Cursor Pagination - uses (rankScore, _id) compound key
-   */
-  /**
-   * Cursor-based pagination for ranked feed
-   * @recommended Preferred over getRankedFeed for production infinite scroll
    * @pattern Cursor Pagination with Weighted Scoring
-   * @complexity O(1) cursor lookup vs O(skip) scan; scores computed once before cursor filtering
+   * @complexity scores computed once before cursor filtering
    * @performance Handles deep pagination efficiently without O(skip) overhead
    * @param favoriteTags - Tag names for tag-match scoring component
    * @param options - Cursor navigation with custom score weights
@@ -988,7 +854,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
       }
 
       const sortDirection = direction === "forward" ? -1 : 1;
-      const feedProjection = this.getStandardProjectionFields();
+      const feedProjection = getStandardProjectionFields();
 
       const pipeline: PipelineStage[] = [
         { $match: withActivePostFilter({ createdAt: { $gte: sinceDate } }) },
@@ -1039,7 +905,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
         { $sort: { rankScore: sortDirection, _id: sortDirection } },
         { $limit: limit + 1 },
         // populate relationships only on paginated results
-        ...this.getStandardLookups(),
+        ...getStandardLookups(),
         {
           $project: {
             ...feedProjection,
@@ -1093,202 +959,6 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
     }
   }
 
-  /**
-   * Backward-compatible entry point for callers that still ask for the old
-   * facet variant. The new-feed path intentionally uses the two-query
-   * implementation to avoid MongoDB's 16MB $facet result document limit.
-   */
-  async getNewFeedWithFacet(
-    limit: number,
-    skip: number,
-  ): Promise<PaginationResult<FeedPost>> {
-    return this.getNewFeed(limit, skip);
-  }
-
-  /**
-   * Backward-compatible entry point for callers that still ask for the old
-   * facet variant. This delegates to the two-query implementation to avoid
-   * MongoDB's 16MB $facet result document limit.
-   */
-  async getTrendingFeedWithFacet(
-    limit: number,
-    skip: number,
-    options?: {
-      timeWindowDays?: number;
-      minLikes?: number;
-      weights?: { recency?: number; popularity?: number; comments?: number };
-    },
-  ): Promise<PaginationResult<FeedPost>> {
-    return this.getTrendingFeed(limit, skip, options);
-  }
-
-  /**
-   * returns standard $lookup stages for populating post relationships
-   */
-
-  private getStandardLookups(): PipelineStage.FacetPipelineStage[] {
-    return [
-      {
-        $lookup: {
-          from: "tags",
-          localField: "tags",
-          foreignField: "_id",
-          as: "tagObjects",
-        },
-      },
-      {
-        $lookup: {
-          from: "images",
-          localField: "image",
-          foreignField: "_id",
-          as: "imageDoc",
-        },
-      },
-      { $unwind: { path: "$imageDoc", preserveNullAndEmptyArrays: true } },
-
-      // lookup community for community posts
-      {
-        $lookup: {
-          from: "communities",
-          localField: "communityId",
-          foreignField: "_id",
-          as: "communityDoc",
-        },
-      },
-      { $unwind: { path: "$communityDoc", preserveNullAndEmptyArrays: true } },
-
-      {
-        $lookup: {
-          from: "posts",
-          let: { repostId: "$repostOf" },
-          pipeline: [
-            {
-              $match: {
-                $expr: { $eq: ["$_id", "$$repostId"] },
-                ...ACTIVE_POST_FILTER,
-              },
-            },
-
-            {
-              $lookup: {
-                from: "images",
-                localField: "image",
-                foreignField: "_id",
-                as: "repostImageDoc",
-              },
-            },
-            {
-              $unwind: {
-                path: "$repostImageDoc",
-                preserveNullAndEmptyArrays: true,
-              },
-            },
-          ],
-          as: "repostDoc",
-        },
-      },
-      { $unwind: { path: "$repostDoc", preserveNullAndEmptyArrays: true } },
-    ];
-  }
-
-  /**
-   * returns standard projection fields for shaping post output
-   */
-
-  private getStandardProjectionFields(): Record<string, unknown> {
-    return {
-      _id: 0,
-      publicId: 1,
-      body: 1,
-      slug: 1,
-      type: 1,
-      repostCount: 1,
-      createdAt: 1,
-      likes: "$likesCount",
-      viewsCount: { $ifNull: ["$viewsCount", 0] },
-      commentsCount: 1,
-
-      // Map the root userPublicId directly to the author snapshot
-      userPublicId: "$author.publicId",
-
-      tags: {
-        $map: {
-          input: { $ifNull: ["$tagObjects", []] },
-          as: "tag",
-          in: { tag: "$$tag.tag", publicId: "$$tag.publicId" },
-        },
-      },
-
-      // Construct the User object from the Snapshot
-      user: {
-        publicId: "$author.publicId",
-        handle: "$author.handle",
-        username: "$author.username",
-        avatar: "$author.avatarUrl",
-        displayName: "$author.displayName",
-      },
-
-      image: {
-        $cond: {
-          if: { $ne: ["$imageDoc", null] },
-          then: {
-            publicId: "$imageDoc.publicId",
-            url: "$imageDoc.url",
-            slug: "$imageDoc.slug",
-          },
-          else: {},
-        },
-      },
-
-      repostOf: {
-        $cond: {
-          if: { $ne: ["$repostDoc", null] },
-          then: {
-            publicId: "$repostDoc.publicId",
-            body: "$repostDoc.body",
-            slug: "$repostDoc.slug",
-            likesCount: "$repostDoc.likesCount",
-            commentsCount: "$repostDoc.commentsCount",
-            repostCount: "$repostDoc.repostCount",
-
-            user: {
-              publicId: "$repostDoc.author.publicId",
-              handle: "$repostDoc.author.handle",
-              username: "$repostDoc.author.username",
-              avatar: "$repostDoc.author.avatarUrl",
-            },
-
-            image: {
-              $cond: {
-                if: { $ne: ["$repostDoc.repostImageDoc", null] },
-                then: {
-                  publicId: "$repostDoc.repostImageDoc.publicId",
-                  url: "$repostDoc.repostImageDoc.url",
-                },
-                else: null,
-              },
-            },
-          },
-          else: null,
-        },
-      },
-
-      // community info for community posts
-      community: {
-        $cond: {
-          if: { $ne: ["$communityDoc", null] },
-          then: {
-            publicId: "$communityDoc.publicId",
-            name: "$communityDoc.name",
-            slug: "$communityDoc.slug",
-            avatar: "$communityDoc.avatar",
-          },
-          else: null,
-        },
-      },
-    };
-  }
-
   private async loadFavoriteTagIds(
     tagNames: string[],
   ): Promise<mongoose.Types.ObjectId[]> {
@@ -1310,7 +980,7 @@ export class FeedReadDao extends BaseRepository<IPost> implements IFeedReadDao {
     const fetchPromise = this.tagRepository
       .findByTags(tagNames)
       .then((tagDocs) =>
-        tagDocs.map((doc) => this.normalizeObjectId(doc._id, "tag._id")),
+        tagDocs.map((doc) => normalizeObjectId(doc._id, "tag._id")),
       )
       .then((ids) => {
         this.tagIdCacheStore.set(cacheKey, {
